Move fetchClusters logic from exporterService to nutanix package
